updates BedrockIT, MixedbreadAIIT, OpenAIAzureIT, OpenAIGroqAPIsIT, OpenAIMistralApiIT, WatsonIT
