+++ extended/src/test/java/apoc/ml/OpenAIGroqAPIsIT.java
@@ -1,59 +1,41 @@
 package apoc.ml;
 
-import apoc.util.TestUtil;
-import org.junit.Assume;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.neo4j.test.rule.DbmsRule;
-import org.neo4j.test.rule.ImpermanentDbmsRule;
+import org.junit.experimental.runners.Enclosed;
+import org.junit.runner.RunWith;
 
-import java.util.Map;
-
-import static apoc.ml.MLUtil.*;
-import static apoc.ml.OpenAITestResultUtils.CHAT_COMPLETION_QUERY;
-import static apoc.ml.OpenAITestResultUtils.assertChatCompletion;
-import static apoc.util.TestUtil.testCall;
-
-
-/**
- * Tests with Groq API: https://console.groq.com/docs/quickstart 
- */
+@RunWith(Enclosed.class)
 public class OpenAIGroqAPIsIT {
 
-    private String groqApiKey;
-
-    @Rule
-    public DbmsRule db = new ImpermanentDbmsRule();
 
 
-    @Before
-    public void setUp() throws Exception {
-        groqApiKey = System.getenv("GROQ_API_KEY");
-        Assume.assumeNotNull("No GROQ_API_KEY environment configured", groqApiKey);
-        TestUtil.registerProcedure(db, OpenAI.class);
+    public static class OpenAIGroqAPIsITllama2 extends OpenAiGroqAPIsBaseIT{
+
+        public static final String MODEL = "llama2-70b-4096";
+
+        @Override
+        String getDefModel() {
+            return MODEL;
+        }
     }
 
-    @Test
-    public void chatCompletionWithLlama2() {
-        String model = "llama2-70b-4096";
-        testCall(db, CHAT_COMPLETION_QUERY,
-                getParams(model),
-                (row) -> assertChatCompletion(row, model));
+    public static class OpenAIGroqAPIsITmixtral extends OpenAiGroqAPIsBaseIT{
+
+        public static final String MODEL = "mixtral-8x7b-32768";
+
+        @Override
+        String getDefModel() {
+            return MODEL;
+        }
     }
 
-    @Test
-    public void chatCompletionWithMixtral() {
-        String model = "mixtral-8x7b-32768";
-        testCall(db, CHAT_COMPLETION_QUERY,
-                getParams(model),
-                (row) -> assertChatCompletion(row, model));
+    public static class OpenAIGroqAPIsITllama3 extends OpenAiGroqAPIsBaseIT{
+
+        public static final String MODEL = "llama-3.3-70b-versatile";
+
+        @Override
+        String getDefModel() {
+            return MODEL;
+        }
     }
-    
-    private Map<String, Object> getParams(String model) {
-        Map<String, String> conf = Map.of(ENDPOINT_CONF_KEY, "https://api.groq.com/openai/v1",
-                MODEL_CONF_KEY, model);
-        
-        return Map.of("apiKey", groqApiKey, "conf", conf);
-    }
+
 }
